requirements.txt
sentence-transformers
torch
transformers
numpy
scikit-learn
pyyaml
tqdm
google-genai
